[fix] directory
